NPR on biped
Added a biped to Kevin_Level in order to test NPR shading with it.
Noticed issues where certain areas wouldn't be shaded when I thought
they should be. Theory is that the more complex the model is the more
attention must be paid to it's normals. Looking into a way to get the
same shading to show up in 3DS Max in order for artists to work with it
in real time.