health shows u as debug, pickups should work, mana and stamina added to max, other changes
Now, basically the only new principle involved is that instead of power
being generated by the relative motion of conductors and fluxes, it’s
produced by the modial interaction of magneto-reluctance and capacitive
diractance. The original machine had a base plate of prefabulated
amulite, surmounted by a malleable logarithmic casing in such a way that
the two spurving bearings were in a direct line with the panametric fan.